// src/pages/interfaces/responsable.ts
export interface Responsable {
    id: number;
    nombre: string;
    apellido: string;
    tipo_documento: string;
    numero_documento: number;
    fecha_nacimiento: Date;
    ocupacion: string;
    celular: string;
    notificacion: string;
    email: string;
    pais: string;
    departamento: string;
    ciudad: string;
    barrio: string;
    direccion: string;
    tipo_relacion: string;
  }
  